+++ RELEASES.md
@@ -37,6 +37,11 @@ Version 1.17.0 (unreleased)
         [`wrapping_div_euclid`][f-wde-1-17],
         [`unwrapped_div_euclid`][f-ude-1-17],
         [`overflowing_div_euclid`][f-ode-1-17]
+      * [`div_euclid_int`][f-dei-1-17], [`checked_div_euclid_int`][f-cdei-1-17],
+        [`saturating_div_euclid_int`][f-sdei-1-17],
+        [`wrapping_div_euclid_int`][f-wdei-1-17],
+        [`unwrapped_div_euclid_int`][f-udei-1-17],
+        [`overflowing_div_euclid_int`][f-odei-1-17]
       * [`lerp`][f-l-1-17], [`checked_lerp`][f-cl-1-17],
         [`saturating_lerp`][f-sl-1-17], [`wrapping_lerp`][f-wl-1-17],
         [`unwrapped_lerp`][f-ul-1-17], [`overflowing_lerp`][f-ol-1-17]
@@ -58,11 +63,13 @@ Version 1.17.0 (unreleased)
 
 [f-cd-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.checked_div
 [f-cde-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.checked_div_euclid
+[f-cdei-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.checked_div_euclid_int
 [f-cil-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.checked_inv_lerp
 [f-cl-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.checked_lerp
 [f-cr-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.checked_recip
 [f-cri-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.checked_rem_int
 [f-de-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.div_euclid
+[f-dei-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.div_euclid_int
 [f-fs-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.from_str
 [f-fsb-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.from_str_binary
 [f-fsh-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.from_str_hex
@@ -71,6 +78,7 @@ Version 1.17.0 (unreleased)
 [f-l-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.lerp
 [f-od-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.overflowing_div
 [f-ode-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.overflowing_div_euclid
+[f-odei-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.overflowing_div_euclid_int
 [f-ofs-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.overflowing_from_str
 [f-ofsb-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.overflowing_from_str_binary
 [f-ofsh-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.overflowing_from_str_hex
@@ -81,6 +89,7 @@ Version 1.17.0 (unreleased)
 [f-r-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.recip
 [f-sd-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.saturating_div
 [f-sde-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.saturating_div_euclid
+[f-sdei-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.saturating_div_euclid_int
 [f-sfs-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.saturating_from_str
 [f-sfsb-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.saturating_from_str_binary
 [f-sfsh-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.saturating_from_str_hex
@@ -90,6 +99,7 @@ Version 1.17.0 (unreleased)
 [f-sr-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.saturating_recip
 [f-ud-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.unwrapped_div
 [f-ude-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.unwrapped_div_euclid
+[f-udei-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.unwrapped_div_euclid_int
 [f-ufs-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.unwrapped_from_str
 [f-ufsb-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.unwrapped_from_str_binary
 [f-ufsh-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.unwrapped_from_str_hex
@@ -100,6 +110,7 @@ Version 1.17.0 (unreleased)
 [f-uri-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.unwrapped_rem_int
 [f-wd-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.wrapping_div
 [f-wde-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.wrapping_div_euclid
+[f-wdei-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.wrapping_div_euclid_int
 [f-wfs-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.wrapping_from_str
 [f-wfsb-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.wrapping_from_str_binary
 [f-wfsh-1-17]: https://docs.rs/fixed/~1.16/fixed/struct.FixedI32.html#method.wrapping_from_str_hex